+++ release_notes.md
@@ -1,5 +1,6 @@
 ## 1.0.0-beta.2
 
+* + lt, lte, gt, substring, ordered in match
 * + orderInteraction; implemented match for all cardinalities
 * + selectPoint, positionObjectInteraction; implemented scoring with baseType point
 * if no itemResult/candidateResponse is provided add an itemResult with outcomes 0 instead of returning an error.